beginning of LSH
