+++ minlizhash/min_hash.py
@@ -76,45 +76,7 @@ def get_signature_matrix(
 
 # some of this was just gpt4, going to rewrite it better tomorrow
 # (to be clear, very, very little over this code generally is gpt4 lol)
-def lsh_buckets_vectorized(
-    signatures: np.ndarray, num_bands: int, seed: int = 42
-) -> DefaultDict[int, Set[int]]:
-    """
-    Organize documents into LSH buckets using vectorized operations.
 
-    :param signatures: Signature matrix (documents x hash functions).
-    :param num_bands: Number of bands for LSH.
-    :param seed: Seed value for hash function.
-    :return: A dictionary mapping each bucket hash to a set of document IDs.
-    """
-    num_docs, num_hashes = signatures.shape
-    assert (
-        num_hashes % num_bands == 0
-    ), "Number of hashes must be divisible by num_bands"
-
-    rows_per_band = num_hashes // num_bands
-    buckets: DefaultDict[int, Set[int]] = defaultdict(set)
-
-    # Precompute all hashes for each band and each document
-    all_band_hashes = np.zeros((num_docs, num_bands), dtype=np.uint64)
-
-    for band in range(num_bands):
-        # Efficiently compute the hashes for each band across all documents
-        start_idx = band * rows_per_band
-        band_signatures = signatures[:, start_idx : start_idx + rows_per_band]
-        # hash_numpy_vec
-        for doc_id, band_signature in enumerate(band_signatures):
-            # Use an efficient hash function such as xxhash
-            all_band_hashes[doc_id, band] = xxh32(
-                np.append(band_signature, seed + band)
-            ).intdigest()
-
-    # Populate the buckets
-    for doc_id in range(num_docs):
-        for band in range(num_bands):
-            buckets[all_band_hashes[doc_id, band]].add(doc_id)
-
-    return buckets
 
 
 def hash_band(band: np.ndarray, seed: int = 42) -> int:
